+++ spec/attachment/AttachmentProcessor.spec.ts
@@ -3,7 +3,7 @@ import chai from 'chai';
 import fs from 'fs';
 import { Readable } from 'stream';
 import sinon from 'sinon';
-import { ObjectStorage } from '@elastic.io/maester-client/dist';
+import { ObjectStorage } from '@elastic.io/maester-client';
 import { AttachmentProcessor, STORAGE_TYPE_PARAMETER, MAESTER_OBJECT_ID_ENDPOINT } from '../../src/attachment/AttachmentProcessor';
 
 const { expect } = chai;
@@ -20,7 +20,6 @@ const formStream = (dataString: string): Readable => {
 
 describe('AttachmentProcessor', () => {
   const attachmentProcessor = new AttachmentProcessor();
-
   describe('Steward', () => {
     it('Should successfully retrieve csv', async () => {
       const attachmentOptions = {
@@ -33,7 +32,6 @@ describe('AttachmentProcessor', () => {
       const expectedResult = fs.readFileSync('spec/attachment/resources/base64csv.txt').toString();
       expect(encodedResult).to.be.equal(expectedResult);
     });
-
     it('Should successfully retrieve png image', async () => {
       const attachmentOptions = {
         'content-type': 'arraybuffer',
@@ -49,7 +47,7 @@ describe('AttachmentProcessor', () => {
   describe('maester', () => {
     let getById;
     beforeEach(() => {
-      getById = sinon.stub(ObjectStorage.prototype, 'getById').callsFake(async () => ({ data: formStream('i`m a stream') }));
+      getById = sinon.stub(ObjectStorage.prototype, 'getOne').callsFake(async () => ({ data: formStream('i`m a stream') }));
     });
     afterEach(() => {
       sinon.restore();
@@ -63,7 +61,7 @@ describe('AttachmentProcessor', () => {
       const result: any = await attachmentProcessor.getAttachment(attachmentOptions.url, attachmentOptions['content-type']);
       expect(result.toString('base64')).to.be.equal({ data: formStream('i`m a stream') }.toString());
       expect(getById.getCall(0).args[0]).to.be.equal('object_id');
-      expect(getById.getCall(0).args[1]).to.be.equal('stream');
+      expect(getById.getCall(0).args[1]).to.be.deep.equal({ responseType: 'stream' });
     });
   });
 });